requirements.txt
streamlit==1.45.1
google-generativeai==0.8.5
python-dotenv==1.1.0
bleach==6.2.0
pytest==8.3.4
pytest-cov==6.1.1
